requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
boto3==1.34.71
botocore==1.34.71
build==1.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-formtools==2.5.1
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-storages==1.14.2
django-two-factor-auth==1.16.0
djangorestframework==3.14.0
drf-nested-routers==0.93.5
idna==3.6
jmespath==1.0.1
morpho-typing==1.5.0
numpy==1.26.4
packaging==24.0
phonenumberslite==8.13.31
psycopg==3.1.19
pydantic==2.6.4
pydantic_core==2.16.3
pypng==0.20220715.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
gunicorn
django-redis
django-cachalot
serpy
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
pyjwt==2.8.0
regex==2024.7.24
